Add the mesh calibration UI
A new mesh calibration flow was added under the build plate settings
menu, including prompts for making sure the build plate is seated
and cleaning the extruders.  The menu button to access this flow
also includes a toggle switch to turn mesh compensation on or off,
and attempting to toggle mesh compensation on while there is no
available mesh data will enter the mesh calibration flow.

The temperature status component has been updated so that specifying
"chamber and build plate" on a machine with no build plate will fall
back to just displaying the build plate.